+++ src/main/java/seng201/team56/models/RoundDifficulty.java
@@ -1,14 +1,14 @@
 package seng201.team56.models;
 
 public enum RoundDifficulty {
-	EASY(1,1.5f),
+	EASY(1,1.5),
 	MEDIUM(2,3),
 	HARD(3,5);
 	
-	private final float minBaseModifier;
-	private final float maxBaseModifier;
+	private final double minBaseModifier;
+	private final double maxBaseModifier;
 	
-	RoundDifficulty(float minBaseModifier, float maxBaseModifier) {
+	RoundDifficulty(double minBaseModifier, double maxBaseModifier) {
 		this.maxBaseModifier = maxBaseModifier;
 		this.minBaseModifier = minBaseModifier;
 	}
